🚧 Move navigateToPracticeTab function to dedicated file, and try harder to make sure page loads before we try to navigate.
